requirements.txt
blinker==1.9.0
click==8.1.8
filelock==3.17.0
Flask==3.1.0
fsspec==2024.12.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
sympy==1.13.3
torch==2.2.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
requests==2.28.0
flask-cors>=3.1.1
transformers==4.26.1
Pillow==9.5.0
numpy==1.23.5
matplotlib==3.6.2



